move custom parsing code of tracing events to the tracing crate
